# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "Travel Stories API",
    "description": "REST API documentation for the travel platform",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Login user",
        "operationId": "loginUser",
        "description": "Authenticate a user and return access and refresh tokens",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/loginUser"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/200-login"
          },
          "401": {
            "$ref": "#/components/responses/401-userNotFound"
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Register a new user",
        "operationId": "registerUser",
        "description": "Create a new user account",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/registerUser"
              }
            }
          }
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/201-register"
          },
          "409": {
            "$ref": "#/components/responses/409"
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Logout user",
        "operationId": "logoutUser",
        "description": "Log out user and invalidate access token",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "$ref": "#/components/responses/204"
          }
        }
      }
    },
...
    "/auth/reset-pwd": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Reset password",
        "operationId": "resetPassword",
        "description": "Reset the user's password using a token from email",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/resetPassword"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/200-resetPwd"
          },
          "404": {
            "$ref": "#/components/responses/404-user"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
...
      "204": {
        "description": "Successfully deleted contact"
      },
...
      "409": {
        "description": "Conflict - Duplicate resource",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 409
                },
                "message": {
                  "type": "string",
                  "example": "Email in use"
                }
              }
            }
          }
        }
      },
...
      "200-login": {
        "description": "Successfully logged in user",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 200
                },
                "message": {
                  "type": "string",
                  "example": "Successfully logged in an user!"
                },
                "data": {
                  "type": "object",
                  "required": [
                    "accessToken"
                  ],
                  "properties": {
                    "accessToken": {
                      "type": "string",
                      "example": "EXelG/M5d9DRudgt0oFOqaiV3uHRbxdz/TltoUrI"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "401-userNotFound": {
        "description": "Unauthorized. User not found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 401
                },
                "message": {
                  "type": "string",
                  "example": "UnauthorizedError"
                },
                "data": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "User not found"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "201-register": {
        "description": "Successfully created a new contact",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 201
                },
                "message": {
                  "type": "string",
                  "example": "Successfully registered a contact!"
                },
                "data": {
                  "$ref": "#/components/schemas/user"
                }
              }
            }
          }
        }
      },
...
      "200-resetPwd": {
        "description": "Successfully reset user password",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 200
                },
                "message": {
                  "type": "string",
                  "example": "Password was successfully reset!"
                }
              }
            }
          }
        }
      },
      "404-user": {
        "description": "Contact not found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 404
                },
                "message": {
                  "type": "string",
                  "example": "User not found"
                },
                "data": {
                  "type": "object",
                  "example": {}
                }
              }
            }
          }
        }
      },
...
    },
    "schemas": {
      "loginUser": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "example": "john@example.com"
          },
          "password": {
            "type": "string",
            "example": "myStrongPassword123"
          }
        }
      },
      "registerUser": {
        "type": "object",
        "required": [
          "name",
          "email",
          "password"
        ],
        "properties": {
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "example": "john@example.com"
          },
          "password": {
            "type": "string",
            "example": "myStrongPassword123"
          }
        }
      },
      "user": {
        "type": "object",
        "required": [
          "_id",
          "name",
          "email",
          "password"
        ],
        "properties": {
          "_id": {
            "description": "User's unique identifier (MongoDB ObjectId)",
            "type": "string",
            "example": "65e4decdd286b30065d54af9"
          },
          "name": {
            "description": "User's full name",
            "type": "string",
            "example": "Serhii Gaievyi"
          },
          "avatarUrl": {
            "description": "Link to user's avatar image",
            "type": "string",
            "format": "uri",
            "example": "https://res.cloudinary.com/demo/image/upload/avatar.png"
          },
          "articlesAmount": {
            "description": "Number of articles created by the user",
            "type": "integer",
            "example": 12
          },
          "description": {
            "description": "Short description or bio of the user",
            "type": "string",
            "example": "Full Stack Developer from Ukraine"
          },
          "email": {
            "description": "User's email address",
            "type": "string",
            "format": "email",
            "example": "gaevuha@gmail.com"
          },
          "password": {
            "description": "User's password (hashed in the database)",
            "type": "string",
            "example": "$2b$10$abcdefgHashedPasswordExample"
          },
          "createdAt": {
            "description": "Date when the user was created",
            "type": "string",
            "format": "date-time",
            "example": "2025-11-06T18:32:41.123Z"
          },
          "updatedAt": {
            "description": "Date when the user was last updated",
            "type": "string",
            "format": "date-time",
            "example": "2025-11-06T18:45:12.456Z"
          }
        }
      },
...
      "resetPassword": {
        "type": "object",
        "required": [
          "token",
          "newPassword"
        ],
        "properties": {
          "token": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "newPassword": {
            "type": "string",
            "example": "NewStrongPassword123"
          }
        }
      },
...
    }
  }
}